Fixed bug in idRenderWorldLocal::TraceXXX.
The idea was that we first check ray-area intersection, and iterate over entities/surfaces of the area only if area is intersected.

However, we don't have proper area bounds anywhere!
Taking bounds of area surfaces is wrong, because area might have some of its boundaries in caulk --- such boundaries are not included.
You can see such a case in test\area_bounds_not_contain_entities.map.

On the other hand, we already do intersection with areas win "fastWorld" case.
In this case, we use BSP tree (the only data which represents full area volume) to intersect with the ray.
So we already know perfectly which areas do intersect the ray, which don't.
We don't even need to intersect with area bounds in this case.

Note: perhaps for fastWorld == false we should still run BSP code, but only use its area set?

This code is used in CPU-side raytracing of render world --- most typical usage is static collisions of particles.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10465 49c82d7f-2e2a-0410-a16f-ae8f201b507f