Fixed occasional floating point overflows in idBoxOctree.
If an object is moved along X axis, then we compute "inverse velocity", which is clamped to 1e+10 for Y and Z.
Also when we traverse octree, we maintain AAbox representing space of the current node, and it is initially from -1e+30 tp 1e+30.
When these two values combine, they generate temporaries up to 1e+40, which does not fit float.

Now aabox of the whole universe is from -1e+10 to 1e+10.
It should be more than enough to cover entire level, and 1e+10 * 1e+10 = 1e+20 is still below FLT_MAX ~= 1e+38.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10491 49c82d7f-2e2a-0410-a16f-ae8f201b507f