Fixed MovementSubsystem::CheckBlocked using uninitialized values on game start.
This code uses ring buffer of values, which includes _originHistory and _frameHistory.
The ring buffers is of fixed size, and there is a pointer moving cyclically though it.

However, ring buffer was not initialized initially.
Now all entries of ring buffer are initialized to time = 0 and origin = current origin.
It works as if we actually had some frames before game start, and AI was standing still during those frames.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10494 49c82d7f-2e2a-0410-a16f-ae8f201b507f