#6333. Disabled zeroing player velocity which broke jumps on high FPS.
I tracked this code to changes in #4493 by BluePill.
I have no idea why this code existed: in my testing player still moves OK.
If we find out why this code is necessary, we can uncomment it and limit it to non-flying cases (walk only perhaps).

Note that there is another similar code for zeroing small velocity at the beginning of Friction.
I think it is enough to avoid small nonzero movements.
Moreover, every such piece of code is dangerous because it might cause such "sticking" issues as we have here.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10499 49c82d7f-2e2a-0410-a16f-ae8f201b507f