Some cleanup in render entity/light update.
Added check member in debug, which we assert for being zero.
This is not very reliable, but it should catch cases when coder creates something like renderEntity_t but forgets to memset it to zero.
And we need zeroing all the memory including padding bits, since otherwise memcmp comparison won't work.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10523 49c82d7f-2e2a-0410-a16f-ae8f201b507f