Fixed hiding virtual method warning for IsMantleable.
Here idAFAttachment has wrong override due to missing const.
So virtual method calls does not reach this method, while programmer most likely expected it does.

The overriden method does the same as the base method, but additionally marks attachment as unmantleable if its carrier is unmantleable.
I think this is a good change (don't see issues with it).
And if we break something here, at least that would be noticed rather quickly.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10591 49c82d7f-2e2a-0410-a16f-ae8f201b507f