Fixed hiding virtual method warning for PlayFootStepSound.
This function normally does not accept any arguments, but idPlayer version does accept two arguments and is this not callable virtually.
As far as I see, Stifu added these arguments rather recently for mantlng changes.

I have renamed this extended method to PlayPlayerFootStepSound, it is not virtual anymore.
Also created proper virtual override PlayFootStepSound, which calls PlayPlayerFootStepSound with default arguments.

This method is normally called only from idPlayer class, in which case it does not matter if virtual chain is broken.
Also this method is called from footstep/leftFoot/rightFoot script events, but they are not called anywhere: neither in core, nor in missions.
So there is no behavior change from fixing this virtual chain.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10593 49c82d7f-2e2a-0410-a16f-ae8f201b507f