#5172, #6321. Use Doom 3 style light frustum for all light culling, never use BFG-style frustum.
These two frustums are slightly different for some spotlights.
Basically:
 1) Doom 3 spotlight can have non-frustum shape (aka deformed frustum), in which case BFG frustum bounds it
 2) BFG frustum must have nonzero "near plane", while Doom 3 frustum can start straight from light origin
See #5815 for in-depth explanation.

In particular, the greatest problem is in p.2: here BFG frustum is LESS than Doom 3 frustum.
So when we use BFG frustum for culling, we can cull away some stuff which we should not.

In my case, moving a spotlight through doorway resulted in temporary absence of many interactions.
Because BFG frustum was already in the next room, which Doom 3 frustum was yet in the current room.
BEtter just use the same frustum everywhere.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10608 49c82d7f-2e2a-0410-a16f-ae8f201b507f