Improved precision of Sys_GetTimeMicroseconds by using newer WinAPI function.
Previously, it used GetSystemTimeAsFileTime.
That function gives precision of about 1 ms.

Now the function uses GetSystemTimePreciseAsFileTime, which seems to provide very good precision.
Unfortunately, it is only available on Windows 8+.
So if someone is playing on Windows 7, this function will still use GetSystemTimeAsFileTime and have 1 ms precision.

Maybe we'll implement some approximation using GetSystemTimeAsFileTime + QueryPerformanceCounter for Windows 7 in the future...

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10658 49c82d7f-2e2a-0410-a16f-ae8f201b507f