Added r_lockView debug cvar.
It allows to move real camera around while keeping frontend camera location fixed.
This is helpful for investigating how frontend works, but be useful for performance analysis of maps.
This new cvar achieves the same as the old cvar r_lockSurfaces was intended to (now it is quite broken).

Note that this debug mode disables some optimizations:
1) r_useScissor / r_useDepthBoundsTest --- everything dependent on it is computed in frontend for specific camera location, it cannot be applies to real camera location when it differs.
2) idFuncPortal controls --- have to block all changes here.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10662 49c82d7f-2e2a-0410-a16f-ae8f201b507f