#6509. Synchronize "persistent level info" between game and main menu GUI.
This data is used in campaigns to pass arbitrary data between missions.
Game scripts already can read and write this persistent info, but it was not accessible in GUI briefings/debriefings before this moment.
Now it is accessible in GUI: each persistent argument "omgName" is "gui::persistent_omgName" in GUI scripts.

The information is copied back and forth between game and main menu.
When game is active, the game is considered source of truth, otherwise main menu GUI is the source of truth (persistentLevelInfoLocation_t says where is truth currently).
The other side may also have some persistent info, but it might be obsolete.
Note that main menu physically does not exist if you are in-game but have not yet opened the menu (i.e. never pressed Escape button).
The only moment when persistent info is fully cleared is when player starts new mission (i.e. when button "Start Mission" is clicked).

Also I deleted the old "gui::startSelect" variable which allowed to specify starting location.
Now it is managed with persistent info, so we should set "gui::persistent_builtin_startMap" instead.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10721 49c82d7f-2e2a-0410-a16f-ae8f201b507f