#5453. Improved reports about cvar mission overrides in console.
If mission override is present, the following two actions show cvar value in magenta color:
1) listcvars {prefix}
2) {prefix} + Tab to autocomplete

There is also 3rd action: {cvar name} + Enter to execute
But it was covered in the previous commit, and shows mission override as extra value.
It is also the most verbose command, because it shows all values: main value, default value, mission override value.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10734 49c82d7f-2e2a-0410-a16f-ae8f201b507f