#5453. Dynamically created cvars are now created with empty value and default.
Previously, if a mapper created new cvar dynamically with setcvar("name", "xxx"), then the new cvar will have main value = "xxx", default value = "xxx", and no mission override.
This is rather strange, because the whole existance of the cvar is in fact mission override!

With this change, the new cvar will have default value = "", and main value = "".
Then it is immediately modified as usual, and it gets mission override = "xxx".
If mission overrides are dropped, cvar resets to empty string.
Ideally it should completely disappear, but I don't want to code this: let's say resetting to empty seems like a good compromise.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10737 49c82d7f-2e2a-0410-a16f-ae8f201b507f