Disabled texture compression for lightFalloffImage, lightAmbientDiffuse, lightAmbientSpecular.
These images are uploaded with DXT1 compression, and it is not possible to force-disable compression in material.
Moreover, these images are small and few, and usually consist of smooth gradients.
DXT1 severely reduces quality of gradients, since its colors are stored with 5-bit components instead of 8-bit.

This change does not increase VRAM consumption because:
  1) such images already take negligible size
  2) R8 is not much larger than DXT1 anyway (most textures are grayscale).

But I think it improves precision of image-based lighting (which uses 32 x 32 cubemaps).


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10873 49c82d7f-2e2a-0410-a16f-ae8f201b507f