#6571. Refactored interaction shaders to accept single "int u_flags" instead of many different boolean flags.
It becomes tedious to pass each flag as separate bool everywhere, and it is hard to add new flags.
It is easier to accumulate flags and pass it as one parameter.
Perhaps the only downside is that it is a bit harder to debug (has to decipher bitflags from RenderDoc).


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10892 49c82d7f-2e2a-0410-a16f-ae8f201b507f