#6571. Supported parallax texture matrix (translate only) + added flag showing if any texture matrix is non-trivial.
I guess it's not hard for GPU to check a uniform flag and skip texcoord transformation entirely if it's not set (which is the common case).
So I expect this might get some performance improvement? Not sure though.

Anyway, now texcoord transformations are also supported on parallax stage, but only work properly on translations.
It turns out that even normal map transformations don't work correctly on anything but translations, and supporting parallax would be even harder.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10893 49c82d7f-2e2a-0410-a16f-ae8f201b507f