Fixed heathaze shaders in GUI overlays with nontrivial resolution scaling.
I tried to use X-ray glasses with heatHaze shader and with r_fboResolution 0.6.
It does not work properly after I rewrote the shaders, because they assume the source texture and the currently rendered framebuffer have same resolution.
This is usually the case, but not the case with X-ray glasses: they are rendered into gui FBO with full resolution, while subview image was generated with scaled resolution.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10933 49c82d7f-2e2a-0410-a16f-ae8f201b507f