#6606. Added "createXrayOverlay" script event.
This event must be used for X-ray glasses/GUIs instead of more generic "createOverlay".
It sets correct layer so that X-ray output is considered to be part of 3d game and happens before tonemap.

X-ray GUI overlay is very special/hacky thing and we must be sure we know when one is being created.
Also, this is a breaking change: we'll have to fix all existing missions that use X-ray glasses.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10967 49c82d7f-2e2a-0410-a16f-ae8f201b507f