Added global defines for engine version.
The defines are:
  TDM_VERSION = 214
  TDM_REVISION = 10984
  TDM_VERSION_FULL = 214010894

The global defines are automatically added to all usages of idParser.
I.e. all source code which runs C preprocessor has these defines present automatically.

These defines can be used as last-resort way to handle compatibility breakage between TDM versions.
If the new version of the engine needs a breaking change, we can adjust the missions so that they work fine both in the new version and in the old (currently default) one.

See more:
  https://forums.thedarkmod.com/index.php?/topic/22690-defines-for-engine-version/

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10984 49c82d7f-2e2a-0410-a16f-ae8f201b507f