#6546. Ignore lighting conditions in idAI::GetObservationPosition.
This code is called when some AIs are actively searching for the player.
And I guess the player also has to be far away / not visible / unreachable...
I'm not completely sure why lighting was taken into account here, but I suppose it is not that necessary.
I see some bits of info in #4347 and #4348, but still not sure.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10998 49c82d7f-2e2a-0410-a16f-ae8f201b507f