#6546,#6623. If spawnarg "les_sample_invisible" = 1, then LightEstimateSystem checks samples with invisible materials as well.
The default behavior is to skip surfaces with invisible materials, to avoid e.g. shadow surfaces and other garbage.
But the absence / blood markers are intended to be completely invisible, but should still have light value computed!
The simplest way to achieve this is to set model which is invisible but still not excluded from LES.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@11002 49c82d7f-2e2a-0410-a16f-ae8f201b507f