Fixed liquid model going crazy after loading savegame.
The model tracks rendertime inside it, and does modelling of the time elapsed since last call when its geometry is requested.
However, the internal state is not tracked in savefile, since it is not game state, and models are normally immutable.
As the result, loading previous save resulted in huge leap back in time, which generated huge lerp coefficient.
That created a heavily deformed surface, way beyond its normal characteristics.

Now if the leap back in time is detected, we simply reset state to zero perturbations.
Note that if leap forward in time happens, it is already clamped at 5 frames.
Of course, the exact state of the surface is not restored, but it looks OK normally.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@11036 49c82d7f-2e2a-0410-a16f-ae8f201b507f