#6608. Added validation of Framebuffer state and fixed various state issues present there.
There are two framebuffer methods which are need to be careful with state: Bind, BindDraw, BlitTo*.
I have added full validation of state at the end of all these methods, including checking that GL state is indeed was it is expected to be.
However, this validation only runs in GL debug context: don't want it to slow down TDM for users.

This validation detected a bug: BindDraw sometimes unexpectedly changes read fbo binding, it happens if fbo needs to be generated.
This caused stencil shadows to glitch around starting/stopping menu, now it should not happen anymore.

Another issue that I see is that BlitTo does not restore draw FBO state in case it was different from read FBO state.
But I have not noticed it caused problems in the wild.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@11045 49c82d7f-2e2a-0410-a16f-ae8f201b507f