Supposedly fixed some one-frame clicker after changing soft shadow quality with stencil shadows.
The problem happens because TiledCustomMipmapStage assumes that its FillFrom is called when proper FBO is bound.
However, after quality changes StencilShadowStage::FillStencilShadowMipmaps calls Shutdown + Init.
And Init changes the bound FBO to "default" for some reason, which has wrong viewport (does not take r_fboResolution into account).

Aside from that, the buggy frames have _white image bound instead of mipmapImage.
That's because since mipmapImage is a scratch image, its GL texture is generated by BindAttachment.
So when we bind it at the start of TiledCustomMipmapStage::FillFrom to TMU 0, we actually don't bind it because its GL texture does not exist yet.
It used to be created later, when the corresponding FBOs were first bound.
Now the FBOs are bound inside Init, making sure the image GL mips are generated right on the start.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@11053 49c82d7f-2e2a-0410-a16f-ae8f201b507f