#5485. Use scissor to copy only portion of screen occupied by the mirror into mirror texture.
Mirror texture uses "screen texture gen", so it must be of full screen size even if there is just one small mirror in the middle.
And while we render only this portion (limited by scissor rect), we used to copy the whole "screen" into the texture.

Now CaptureRenderToImage accepts optional scissor argument, which is passed only during mirror subview generation.
This argument is passed to backend, and is finally handled in FrameBufferManager::CopyRender.
Note that this optimization can be disabled by cvar r_copyRenderUseScissor.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@11061 49c82d7f-2e2a-0410-a16f-ae8f201b507f