#6649. Added r_useNodeCommonChildren kind of optimization to idRenderWorldLocal::BoundsInAreas_r.
This optimization is already used under the same cvar when intersecting a frustum with areas (for locating entities and lights into areas).
Now we do the same for intersecting bounds with areas.
Basically, we can prune away a SBP subtree if it only contains void and the area which we have already added to the output.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@11067 49c82d7f-2e2a-0410-a16f-ae8f201b507f