#6649. Don't use BSP optimization for shadow rays for lights which have origin in the void (in LightQuerySystem).
For a typical parallelSky sun light, the light origin is normally far outside the playing area in the void (it is moved far away for parallel light).
So if we use BSP optimization, shadow ray will hit the surface on the void boundary and will always be considered "shadowed".

Now we don't use BSP, so the shadow ray is simply checked for intersection with all render models in the areas.
ParallelSky light rays come through "portal" brush portals, which have no rendermodels, so this should unblock shadows.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@11069 49c82d7f-2e2a-0410-a16f-ae8f201b507f